refactor: restructure app hub, add test for note
