+++ example-app/app/routing/hub/index.ts
@@ -1,0 +1,2 @@
+export { routes } from './app.routes';
+export * from './app.hub';
