+++ example-app/app/routing/routing.hub.ts
