+++ example-app/app/views/bike/hub/bike.hub.ts
@@ -1,6 +1,6 @@
 import { createFeature, createNote, Root, Slice } from 'lib';
 import { routes } from './bike.routes';
-import { appSlice } from '../../../routing/hub/app.hub';
+import { appSlice } from '../../../routing/hub';
 
 export type BikeNote = Root;
